Moved files and excluded GET request from auth
