--- api/src/main/java/com/buurbak/api/trailers/model/Trailer.java
@@ -1,4 +1,0 @@
-package com.buurbak.api.trailers.model;
-
-public class Trailer {
-}
