--- api/src/main/java/com/buurbak/api/trailers/service/TrailerService.java
@@ -1,4 +1,0 @@
-package com.buurbak.api.trailers.service;
-
-public class TrailerService {
-}
