+++ api/src/main/java/com/buurbak/api/trailers/service/TrailerTypeService.java
@@ -1,8 +1,8 @@
-package com.buurbak.api.trailers.application;
+package com.buurbak.api.trailers.service;
 
 
-import com.buurbak.api.trailers.data.TrailerTypeRepository;
-import com.buurbak.api.trailers.domain.TrailerType;
+import com.buurbak.api.trailers.repository.TrailerTypeRepository;
+import com.buurbak.api.trailers.model.TrailerType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
