--- api/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-#spring.datasource.url=jdbc:postgresql://db:5432/buurbak
-spring.datasource.url=jdbc:postgresql://localhost:5432/buurbak
-spring.datasource.username=buurbak
-spring.datasource.password=buurbak
-spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL95Dialect
-spring.jpa.properties.hibernate.format_sql=true
-server.port=3080
-server.servlet.context-path=/api/v1
